+++ tests/testthat/test-summary_stats.R
@@ -1,11 +1,12 @@
 testthat::test_that("summary_formats works as expected", {
-  result <- summary_formats()
+  result <- summary_formats() %>%
+    unlist() # More compact fruition
   res <- testthat::expect_silent(result)
   testthat::expect_snapshot(res)
 
   result <- summary_formats(type = "counts", include_pval = TRUE)
-  expect_true(all(result[c("n", "count", "n_blq")] == "xx."))
-  expect_identical(result[["pval"]], "x.xxxx | (<0.0001)")
+  testthat::expect_true(all(result[c("n", "count", "n_blq")] == "xx."))
+  testthat::expect_identical(result[["pval_counts"]], "x.xxxx | (<0.0001)")
   expect_identical(result[["count_fraction"]], format_count_fraction)
 })
 
@@ -20,15 +21,15 @@ testthat::test_that("summary_labels works as expected", {
 })
 
 testthat::test_that("summary_custom works as expected", {
-  result <- summary_custom()
-  res <- testthat::expect_silent(result)
+  testthat::expect_warning(res <- summary_custom())
+  res$formats <- unlist(res$formats)
   testthat::expect_snapshot(res)
 
-  result <- summary_custom(
+  testthat::expect_warning(res <- summary_custom(
     type = "counts", stats_custom = c("n", "count"),
     formats_custom = c(n = "xx.xx"), labels_custom = c(count = "#"), indent_mods_custom = 2L
-  )
-  res <- testthat::expect_silent(result)
+  ))
+
   testthat::expect_snapshot(res)
 })