+++ NEWS.md
@@ -1,4 +1,15 @@
 # tern 0.9.0.9000
+### New Features
+* Added vectors of default formats and labels in `tern_default_formats` and `tern_default_labels`, respectively.
+* Added getter for statistical method groups with `get_stats`.
+* Added getters for formats `get_format_from_stats()` and labels `get_labels_from_stats()`.
+
+### Enhancements
+* Grouping of valid methods names and their relative default formats and labels into new source file `utils_defaults_handling.R`.
+
+### Miscellaneous
+* Started deprecation of `summary_custom()` and `a_summary()` as a `S3` method.
+* Statistical method for finding `p-value` in the discrete case is now termed `pval_counts`.
 
 # tern 0.9.0
 ### New Features